+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Opening quick edit inlay with one already open causes un-closable inlay
 - Memory leak detected warning in UserDirectiveTextArea
 - Always allow audio capture while testing config
+- EDT warning in Deepgram provider
 
 ## [2024.11.0] - 2024-09-11