+++ CHANGELOG.md
@@ -9,8 +9,8 @@
 - Opening quick edit inlay with one already open causes un-closable inlay
 - Memory leak detected warning in UserDirectiveTextArea
 - Always allow audio capture while testing config
-- EDT warning in Deepgram provider
 - No such element exception when using non-existent language model
+- EDT threading issues
 
 ## [2024.11.0] - 2024-09-11